helper function and variables moved to a separate utils.py file
